requirements.txt
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
pdfplumber>=0.11.0
cbbpy>=1.2.0
boto3>=1.34.0
fastmcp>=2.0.0
httpx>=0.27.0
linkup-sdk>=0.2.0
pydantic>=2.4.0,<3.0.0
